// src/interfaces/DetailsListBasicExampleState.type.ts
import { IDetailsListBasicExampleItem } from "./DetailsListBasicExampleItem.type";

export interface IDetailsListBasicExampleState {
  items: IDetailsListBasicExampleItem[];
  selectionDetails: any;
  isOpen: boolean;
  title: any;
  entity: any;
  Uentity: any;
  GxPSuppliers: any;
  NatureofProduct: any;
  HCPPAGPOReportable: any;
  HealthcareOrganization: any;
  Candidatereimbursement: any;
  PaymentMethod: any;
  TermsofPayment: any;
  BillingCurrency: any;
  TaxStatus: any;
  TaxID: any;
  ContactName: any;
  ContactEmail: any;
  PhoneNumber: any;
  FaxNumber: any;
  Vendorwebsite: any;
  uID: number;
  utitle: string;
  uEntity: any;
  uGxPSuppliers: any;
  uNatureofProduct: any;
  uHCPPAGPOReportable: any;
  uHealthcareOrganization: any;
  uCandidatereimbursement: any;
  uPaymentMethod: any;
  uTermsofPayment: any;
  uTaxStatus: any;
  uTaxID: any;
  uContactName: any;
  uContactEmail: any;
  uPhoneNumber: any;
  uFaxNumber: any;
  isFormVisible: any;
  readOnl: boolean;
  isDetailsFormAvailable: any;
  gdefault: any;
  dataid: number;
  filetitle: any;
  filelink: any;
  fileID: number;
  FileaID: any;
  FilebID: any;
  FilecID: any;
  listitemidA: any;
  listitemidB: any;
  listitemidC: any;
  showCancelIcon: any;
  closeNewForm: any;
  AddressOne: any;
  AddressTwo: any;
  AddressThree: any;
  aCountry: any;
  aState: any;
  aCity: any;
  aZip: any;
  VlistID: any;
  FAddress: any;
  vID: any;
  ShippingAddressOne: any;
  ShippingAddressTwo: any;
  ShippingAddressThree: any;
  ShippingCountry: any;
  ShippingState: any;
  ShippingCity: any;
  ShippingZip: any;
  tagdata: any;
  tagdatavalue: any;
  flag: any;
  Shippingaddressform: any;
  Addresstype: any;
  ShippingVendorCountry: any;
  isError: boolean;
  titleUnique: string;
  taxIDUnique: any;
  isTaxIDError: boolean;
  isEmailError: boolean;
  isPhoneError: boolean;
  spinLoader: boolean;
  spinfomodel: boolean;
  spinfo: any;
  HCPPAGPOName: any;
  HCPPAGPOCountry: any;
  HideHCPCountryNameOptions: boolean;
  SiretCode: any;
  VendorDocumentView: boolean;
  FileEncodedAbsUrlA: any;
  FileLinkFilenameA: any;
  FileEncodedAbsUrlB: any;
  FileLinkFilenameB: any;
  FileEncodedAbsUrlC: any;
  FileLinkFilenameC: any;
  RequestorName: any;
  currentUser: any;
  SectionData: any;
  NotApplication: any;
  DefaultNotApplication: any;
}